requirements.txt
streamlit==1.41.1
streamlit-audiorecorder==0.0.6
audiocraft==1.3.0
matplotlib==3.7.0
python-dotenv==1.0.1